(fga-eps-mds/2023-1-CAPJu-Doc#68) - Adiciona novo teste de formato de data
