+++ src/components/Flow/index.tsx
@@ -11,6 +11,7 @@ import "reactflow/dist/style.css";
 import _ from "lodash";
 
 import { colors } from "styles/colors";
+import handleDateFormating from "utils/dates";
 
 interface FlowProps {
   stages: Stage[];
@@ -98,14 +99,14 @@ export function Flow({
     });
   };
 
-  const handleDateFormating = (date: Date | string) => {
-    const currentDate = new Date(date);
-    return currentDate.toLocaleString("pt-BR", {
-      year: "numeric",
-      month: "long",
-      day: "numeric",
-    });
-  };
+  // const handleDateFormating = (date: Date | string) => {
+  //   const currentDate = new Date(date);
+  //   return currentDate.toLocaleString("pt-BR", {
+  //     year: "numeric",
+  //     month: "long",
+  //     day: "numeric",
+  //   });
+  // };
 
   const handleExpiration = (vencimento: Date) => {
     const currentDate = new Date();
