+++ src/utils/dates.ts
@@ -1,0 +1,10 @@
+const handleDateFormating = (date: Date | string) => {
+  const currentDate = new Date(date);
+  return currentDate.toLocaleString("pt-BR", {
+    year: "numeric",
+    month: "long",
+    day: "numeric",
+  });
+};
+
+export default handleDateFormating;
